refactor: http메소드 refresh 로직 적용
